refactor(eajs): move embed settings types from setup into the script contract
SdkIframeDashboardEmbedSettings and SdkIframeQuestionEmbedSettings move
into script/types/embed.ts and setup re-exports them. Kills the
setup <-> script type cycle: the contract no longer imports from setup.
