+++ frontend/src/metabase/embedding/embedding-iframe-sdk-setup/types.ts
@@ -1,11 +1,11 @@
 import type {
   BrowserEmbedOptions,
-  DashboardEmbedOptions,
   ExplorationEmbedOptions,
   MetabotEmbedOptions,
-  QuestionEmbedOptions,
+  SdkIframeDashboardEmbedSettings,
   SdkIframeEmbedAuthTypeSettings,
   SdkIframeEmbedBaseSettings,
+  SdkIframeQuestionEmbedSettings,
 } from "metabase/embedding/embedding-iframe-sdk/script/types/embed";
 import type {
   BaseRecentItem,
@@ -34,13 +34,10 @@ export type SdkIframeEmbedSetupRecentItem = Pick<
 export type SdkIframeEmbedSetupGuestEmbedSettings =
   SdkIframeEmbedAuthTypeSettings;
 
-export type SdkIframeDashboardEmbedSettings = DashboardEmbedOptions & {
-  lockedParameters?: string[];
-};
-
-export type SdkIframeQuestionEmbedSettings = QuestionEmbedOptions & {
-  lockedParameters?: string[];
-};
+export type {
+  SdkIframeDashboardEmbedSettings,
+  SdkIframeQuestionEmbedSettings,
+} from "metabase/embedding/embedding-iframe-sdk/script/types/embed";
 
 export type SdkIframeEmbedSetupTemplateSettings =
   | SdkIframeDashboardEmbedSettings
